Added mstl_impl module
