+++ src/lib.rs
@@ -4,6 +4,7 @@
 
 mod error;
 mod mstl;
+mod mstl_impl;
 mod mstl_params;
 mod mstl_result;
 mod stl;
